+++ packages/@coorpacademy-components/src/atom/button-link/test/fixtures/button-primary-no-icon.js
