+++ packages/@coorpacademy-components/src/atom/button-link/test/fixtures/button-tertiary-icon-left.js
