+++ packages/@coorpacademy-components/src/atom/button-link/test/fixtures/link-secondary-icon-right.js
