+++ packages/@coorpacademy-components/storybook/components.js
@@ -3,12 +3,12 @@
 import AtomAutocomplete from './../src/atom/autocomplete';
 import AtomAvatar from './../src/atom/avatar';
 import AtomButton from './../src/atom/button';
+import AtomButtonLink from './../src/atom/button-link';
 import AtomCatalogSection from './../src/atom/catalog-section';
 import AtomCenteredText from './../src/atom/centered-text';
 import AtomCheckbox from './../src/atom/checkbox';
 import AtomChips from './../src/atom/chips';
 import AtomClue from './../src/atom/clue';
-import AtomCmButton from './../src/atom/cm-button';
 import AtomContentBadge from './../src/atom/content-badge';
 import AtomCta from './../src/atom/cta';
 import AtomDifficultyLevel from './../src/atom/difficulty-level';
@@ -192,6 +192,12 @@ import AtomButtonFixtureDisabled from '../src/atom/button/test/fixtures/disabled
 import AtomButtonFixtureDownload from '../src/atom/button/test/fixtures/download';
 import AtomButtonFixtureLink from '../src/atom/button/test/fixtures/link';
 import AtomButtonFixtureSso from '../src/atom/button/test/fixtures/sso';
+import AtomButtonLinkFixtureButtonPrimaryIconLeft from '../src/atom/button-link/test/fixtures/button-primary-icon-left';
+import AtomButtonLinkFixtureButtonPrimaryNoIcon from '../src/atom/button-link/test/fixtures/button-primary-no-icon';
+import AtomButtonLinkFixtureButtonTertiaryIconLeft from '../src/atom/button-link/test/fixtures/button-tertiary-icon-left';
+import AtomButtonLinkFixtureIconButton from '../src/atom/button-link/test/fixtures/icon-button';
+import AtomButtonLinkFixtureLinkSecondaryIconRight from '../src/atom/button-link/test/fixtures/link-secondary-icon-right';
+import AtomButtonLinkFixtureLinkTextIconRight from '../src/atom/button-link/test/fixtures/link-text-icon-right';
 import AtomCatalogSectionFixtureList from '../src/atom/catalog-section/test/fixtures/list';
 import AtomCatalogSectionFixturePicture from '../src/atom/catalog-section/test/fixtures/picture';
 import AtomCenteredTextFixtureDefault from '../src/atom/centered-text/test/fixtures/default';
@@ -204,12 +210,6 @@ import AtomChipsFixtureNonSelected from '../src/atom/chips/test/fixtures/non-sel
 import AtomClueFixtureDefault from '../src/atom/clue/test/fixtures/default';
 import AtomClueFixtureLink from '../src/atom/clue/test/fixtures/link';
 import AtomClueFixtureLoading from '../src/atom/clue/test/fixtures/loading';
-import AtomCmButtonFixtureButtonPrimaryIconLeft from '../src/atom/cm-button/test/fixtures/button-primary-icon-left';
-import AtomCmButtonFixtureButtonPrimaryNoIcon from '../src/atom/cm-button/test/fixtures/button-primary-no-icon';
-import AtomCmButtonFixtureButtonTertiaryIconLeft from '../src/atom/cm-button/test/fixtures/button-tertiary-icon-left';
-import AtomCmButtonFixtureIconButton from '../src/atom/cm-button/test/fixtures/icon-button';
-import AtomCmButtonFixtureLinkSecondaryIconRight from '../src/atom/cm-button/test/fixtures/link-secondary-icon-right';
-import AtomCmButtonFixtureLinkTextIconRight from '../src/atom/cm-button/test/fixtures/link-text-icon-right';
 import AtomContentBadgeFixtureAdvanced from '../src/atom/content-badge/test/fixtures/advanced';
 import AtomContentBadgeFixtureArticle from '../src/atom/content-badge/test/fixtures/article';
 import AtomContentBadgeFixtureChapter from '../src/atom/content-badge/test/fixtures/chapter';
@@ -1020,12 +1020,12 @@ export const components = {
     AtomAutocomplete,
     AtomAvatar,
     AtomButton,
+    AtomButtonLink,
     AtomCatalogSection,
     AtomCenteredText,
     AtomCheckbox,
     AtomChips,
     AtomClue,
-    AtomCmButton,
     AtomContentBadge,
     AtomCta,
     AtomDifficultyLevel,
@@ -1252,6 +1252,14 @@ export const fixtures = {
       Link: AtomButtonFixtureLink,
       Sso: AtomButtonFixtureSso
     },
+    AtomButtonLink: {
+      ButtonPrimaryIconLeft: AtomButtonLinkFixtureButtonPrimaryIconLeft,
+      ButtonPrimaryNoIcon: AtomButtonLinkFixtureButtonPrimaryNoIcon,
+      ButtonTertiaryIconLeft: AtomButtonLinkFixtureButtonTertiaryIconLeft,
+      IconButton: AtomButtonLinkFixtureIconButton,
+      LinkSecondaryIconRight: AtomButtonLinkFixtureLinkSecondaryIconRight,
+      LinkTextIconRight: AtomButtonLinkFixtureLinkTextIconRight
+    },
     AtomCatalogSection: {
       List: AtomCatalogSectionFixtureList,
       Picture: AtomCatalogSectionFixturePicture
@@ -1273,14 +1281,6 @@ export const fixtures = {
       Default: AtomClueFixtureDefault,
       Link: AtomClueFixtureLink,
       Loading: AtomClueFixtureLoading
-    },
-    AtomCmButton: {
-      ButtonPrimaryIconLeft: AtomCmButtonFixtureButtonPrimaryIconLeft,
-      ButtonPrimaryNoIcon: AtomCmButtonFixtureButtonPrimaryNoIcon,
-      ButtonTertiaryIconLeft: AtomCmButtonFixtureButtonTertiaryIconLeft,
-      IconButton: AtomCmButtonFixtureIconButton,
-      LinkSecondaryIconRight: AtomCmButtonFixtureLinkSecondaryIconRight,
-      LinkTextIconRight: AtomCmButtonFixtureLinkTextIconRight
     },
     AtomContentBadge: {
       Advanced: AtomContentBadgeFixtureAdvanced,
